Move some shoot-specific utils from common to pkg/utils/gardener
