+++ pkg/operation/common/utils_test.go
@@ -22,13 +22,10 @@ import (
 	"strings"
 	"time"
 
-	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	mockclient "github.com/gardener/gardener/pkg/mock/controller-runtime/client"
 	mocktime "github.com/gardener/gardener/pkg/mock/go/time"
 	. "github.com/gardener/gardener/pkg/operation/common"
-	"github.com/gardener/gardener/pkg/utils"
-	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
 	"github.com/gardener/gardener/pkg/utils/test"
 
 	hvpav1alpha1 "github.com/gardener/hvpa-controller/api/v1alpha1"
@@ -44,7 +41,6 @@ import (
 	schedulingv1 "k8s.io/api/scheduling/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/component-base/version"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -247,226 +243,6 @@ var _ = Describe("common", func() {
 			DNSIncludeZones: "a,b,c",
 			DNSExcludeZones: "d,e,f",
 		}, Equal("bar"), Equal("foo"), Equal([]string{"a", "b", "c"}), Equal([]string{"d", "e", "f"}), Not(HaveOccurred())),
-	)
-
-	DescribeTable("#RespectSyncPeriodOverwrite",
-		func(respectSyncPeriodOverwrite bool, shoot *gardencorev1beta1.Shoot, match gomegatypes.GomegaMatcher) {
-			Expect(RespectShootSyncPeriodOverwrite(respectSyncPeriodOverwrite, shoot)).To(match)
-		},
-
-		Entry("respect overwrite",
-			true,
-			&gardencorev1beta1.Shoot{},
-			BeTrue()),
-		Entry("don't respect overwrite",
-			false,
-			&gardencorev1beta1.Shoot{},
-			BeFalse()),
-		Entry("don't respect overwrite but garden namespace",
-			false,
-			&gardencorev1beta1.Shoot{ObjectMeta: kutil.ObjectMeta(v1beta1constants.GardenNamespace, "foo")},
-			BeTrue()),
-	)
-
-	DescribeTable("#ShouldIgnoreShoot",
-		func(respectSyncPeriodOverwrite bool, shoot *gardencorev1beta1.Shoot, match gomegatypes.GomegaMatcher) {
-			Expect(ShouldIgnoreShoot(respectSyncPeriodOverwrite, shoot)).To(match)
-		},
-
-		Entry("respect overwrite with annotation",
-			true,
-			&gardencorev1beta1.Shoot{ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{ShootIgnore: "true"}}},
-			BeTrue()),
-		Entry("respect overwrite with wrong annotation",
-			true,
-			&gardencorev1beta1.Shoot{ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{ShootIgnore: "foo"}}},
-			BeFalse()),
-		Entry("respect overwrite with no annotation",
-			true,
-			&gardencorev1beta1.Shoot{},
-			BeFalse()),
-	)
-
-	DescribeTable("#IsShootFailed",
-		func(shoot *gardencorev1beta1.Shoot, match gomegatypes.GomegaMatcher) {
-			Expect(IsShootFailed(shoot)).To(match)
-		},
-
-		Entry("no last operation",
-			&gardencorev1beta1.Shoot{},
-			BeFalse()),
-		Entry("with last operation but not in failed state",
-			&gardencorev1beta1.Shoot{
-				Status: gardencorev1beta1.ShootStatus{
-					LastOperation: &gardencorev1beta1.LastOperation{
-						State: gardencorev1beta1.LastOperationStateSucceeded,
-					},
-				},
-			},
-			BeFalse()),
-		Entry("with last operation in failed state but not at latest generation",
-			&gardencorev1beta1.Shoot{
-				ObjectMeta: metav1.ObjectMeta{Generation: 1},
-				Status: gardencorev1beta1.ShootStatus{
-					LastOperation: &gardencorev1beta1.LastOperation{
-						State: gardencorev1beta1.LastOperationStateFailed,
-					},
-				},
-			},
-			BeFalse()),
-		Entry("with last operation in failed state and matching generation but not latest gardener version",
-			&gardencorev1beta1.Shoot{
-				Status: gardencorev1beta1.ShootStatus{
-					LastOperation: &gardencorev1beta1.LastOperation{
-						State: gardencorev1beta1.LastOperationStateFailed,
-					},
-					Gardener: gardencorev1beta1.Gardener{
-						Version: version.Get().GitVersion + "foo",
-					},
-				},
-			},
-			BeFalse()),
-		Entry("with last operation in failed state and matching generation and latest gardener version",
-			&gardencorev1beta1.Shoot{
-				Status: gardencorev1beta1.ShootStatus{
-					LastOperation: &gardencorev1beta1.LastOperation{
-						State: gardencorev1beta1.LastOperationStateFailed,
-					},
-					Gardener: gardencorev1beta1.Gardener{
-						Version: version.Get().GitVersion,
-					},
-				},
-			},
-			BeTrue()),
-	)
-
-	DescribeTable("#IsObservedAtLatestGenerationAndSucceeded",
-		func(shoot *gardencorev1beta1.Shoot, match gomegatypes.GomegaMatcher) {
-			Expect(IsObservedAtLatestGenerationAndSucceeded(shoot)).To(match)
-		},
-
-		Entry("not at observed generation",
-			&gardencorev1beta1.Shoot{
-				ObjectMeta: metav1.ObjectMeta{Generation: 1},
-			},
-			BeFalse()),
-		Entry("last operation state not succeeded",
-			&gardencorev1beta1.Shoot{
-				Status: gardencorev1beta1.ShootStatus{
-					LastOperation: &gardencorev1beta1.LastOperation{
-						State: gardencorev1beta1.LastOperationStateError,
-					},
-				},
-			},
-			BeFalse()),
-		Entry("observed at latest generation and no last operation state",
-			&gardencorev1beta1.Shoot{},
-			BeFalse()),
-		Entry("observed at latest generation and last operation state succeeded",
-			&gardencorev1beta1.Shoot{
-				Status: gardencorev1beta1.ShootStatus{
-					LastOperation: &gardencorev1beta1.LastOperation{
-						State: gardencorev1beta1.LastOperationStateSucceeded,
-					},
-				},
-			},
-			BeTrue()),
-	)
-
-	DescribeTable("#SyncPeriodOfShoot",
-		func(respectSyncPeriodOverwrite bool, defaultMinSyncPeriod time.Duration, shoot *gardencorev1beta1.Shoot, expected time.Duration) {
-			Expect(SyncPeriodOfShoot(respectSyncPeriodOverwrite, defaultMinSyncPeriod, shoot)).To(Equal(expected))
-		},
-
-		Entry("don't respect overwrite",
-			false,
-			1*time.Second,
-			&gardencorev1beta1.Shoot{},
-			1*time.Second),
-		Entry("respect overwrite but no overwrite",
-			true,
-			1*time.Second,
-			&gardencorev1beta1.Shoot{},
-			1*time.Second),
-		Entry("respect overwrite but overwrite invalid",
-			true,
-			1*time.Second,
-			&gardencorev1beta1.Shoot{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{ShootSyncPeriod: "foo"},
-				},
-			},
-			1*time.Second),
-		Entry("respect overwrite but overwrite too short",
-			true,
-			2*time.Second,
-			&gardencorev1beta1.Shoot{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{ShootSyncPeriod: (1 * time.Second).String()},
-				},
-			},
-			2*time.Second),
-		Entry("respect overwrite with longer overwrite",
-			true,
-			2*time.Second,
-			&gardencorev1beta1.Shoot{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{ShootSyncPeriod: (3 * time.Second).String()},
-				},
-			},
-			3*time.Second),
-	)
-
-	Describe("#EffectiveMaintenanceTimeWindow", func() {
-		It("should shorten the end of the time window by 15 minutes", func() {
-			var (
-				begin = utils.NewMaintenanceTime(0, 0, 0)
-				end   = utils.NewMaintenanceTime(1, 0, 0)
-			)
-
-			Expect(EffectiveMaintenanceTimeWindow(utils.NewMaintenanceTimeWindow(begin, end))).
-				To(Equal(utils.NewMaintenanceTimeWindow(begin, utils.NewMaintenanceTime(0, 45, 0))))
-		})
-	})
-
-	DescribeTable("#EffectiveShootMaintenanceTimeWindow",
-		func(shoot *gardencorev1beta1.Shoot, window *utils.MaintenanceTimeWindow) {
-			Expect(EffectiveShootMaintenanceTimeWindow(shoot)).To(Equal(window))
-		},
-
-		Entry("no maintenance section",
-			&gardencorev1beta1.Shoot{},
-			utils.AlwaysTimeWindow),
-		Entry("no time window",
-			&gardencorev1beta1.Shoot{
-				Spec: gardencorev1beta1.ShootSpec{
-					Maintenance: &gardencorev1beta1.Maintenance{},
-				},
-			},
-			utils.AlwaysTimeWindow),
-		Entry("invalid time window",
-			&gardencorev1beta1.Shoot{
-				Spec: gardencorev1beta1.ShootSpec{
-					Maintenance: &gardencorev1beta1.Maintenance{
-						TimeWindow: &gardencorev1beta1.MaintenanceTimeWindow{},
-					},
-				},
-			},
-			utils.AlwaysTimeWindow),
-		Entry("valid time window",
-			&gardencorev1beta1.Shoot{
-				Spec: gardencorev1beta1.ShootSpec{
-					Maintenance: &gardencorev1beta1.Maintenance{
-						TimeWindow: &gardencorev1beta1.MaintenanceTimeWindow{
-							Begin: "010000+0000",
-							End:   "020000+0000",
-						},
-					},
-				},
-			},
-			utils.NewMaintenanceTimeWindow(
-				utils.NewMaintenanceTime(1, 0, 0),
-				utils.NewMaintenanceTime(1, 45, 0))),
 	)
 
 	Describe("#CheckIfDeletionIsConfirmed", func() {
