+++ pkg/utils/gardener/shoot_test.go
@@ -1,0 +1,254 @@
+// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package gardener_test
+
+import (
+	"time"
+
+	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
+	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
+	"github.com/gardener/gardener/pkg/utils"
+	. "github.com/gardener/gardener/pkg/utils/gardener"
+	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/extensions/table"
+	. "github.com/onsi/gomega"
+	gomegatypes "github.com/onsi/gomega/types"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/component-base/version"
+)
+
+var _ = Describe("Shoot", func() {
+	DescribeTable("#RespectSyncPeriodOverwrite",
+		func(respectSyncPeriodOverwrite bool, shoot *gardencorev1beta1.Shoot, match gomegatypes.GomegaMatcher) {
+			Expect(RespectShootSyncPeriodOverwrite(respectSyncPeriodOverwrite, shoot)).To(match)
+		},
+
+		Entry("respect overwrite",
+			true,
+			&gardencorev1beta1.Shoot{},
+			BeTrue()),
+		Entry("don't respect overwrite",
+			false,
+			&gardencorev1beta1.Shoot{},
+			BeFalse()),
+		Entry("don't respect overwrite but garden namespace",
+			false,
+			&gardencorev1beta1.Shoot{ObjectMeta: kutil.ObjectMeta(v1beta1constants.GardenNamespace, "foo")},
+			BeTrue()),
+	)
+
+	DescribeTable("#ShouldIgnoreShoot",
+		func(respectSyncPeriodOverwrite bool, shoot *gardencorev1beta1.Shoot, match gomegatypes.GomegaMatcher) {
+			Expect(ShouldIgnoreShoot(respectSyncPeriodOverwrite, shoot)).To(match)
+		},
+
+		Entry("respect overwrite with annotation",
+			true,
+			&gardencorev1beta1.Shoot{ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{v1beta1constants.ShootIgnore: "true"}}},
+			BeTrue()),
+		Entry("respect overwrite with wrong annotation",
+			true,
+			&gardencorev1beta1.Shoot{ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{v1beta1constants.ShootIgnore: "foo"}}},
+			BeFalse()),
+		Entry("respect overwrite with no annotation",
+			true,
+			&gardencorev1beta1.Shoot{},
+			BeFalse()),
+	)
+
+	DescribeTable("#IsShootFailed",
+		func(shoot *gardencorev1beta1.Shoot, match gomegatypes.GomegaMatcher) {
+			Expect(IsShootFailed(shoot)).To(match)
+		},
+
+		Entry("no last operation",
+			&gardencorev1beta1.Shoot{},
+			BeFalse()),
+		Entry("with last operation but not in failed state",
+			&gardencorev1beta1.Shoot{
+				Status: gardencorev1beta1.ShootStatus{
+					LastOperation: &gardencorev1beta1.LastOperation{
+						State: gardencorev1beta1.LastOperationStateSucceeded,
+					},
+				},
+			},
+			BeFalse()),
+		Entry("with last operation in failed state but not at latest generation",
+			&gardencorev1beta1.Shoot{
+				ObjectMeta: metav1.ObjectMeta{Generation: 1},
+				Status: gardencorev1beta1.ShootStatus{
+					LastOperation: &gardencorev1beta1.LastOperation{
+						State: gardencorev1beta1.LastOperationStateFailed,
+					},
+				},
+			},
+			BeFalse()),
+		Entry("with last operation in failed state and matching generation but not latest gardener version",
+			&gardencorev1beta1.Shoot{
+				Status: gardencorev1beta1.ShootStatus{
+					LastOperation: &gardencorev1beta1.LastOperation{
+						State: gardencorev1beta1.LastOperationStateFailed,
+					},
+					Gardener: gardencorev1beta1.Gardener{
+						Version: version.Get().GitVersion + "foo",
+					},
+				},
+			},
+			BeFalse()),
+		Entry("with last operation in failed state and matching generation and latest gardener version",
+			&gardencorev1beta1.Shoot{
+				Status: gardencorev1beta1.ShootStatus{
+					LastOperation: &gardencorev1beta1.LastOperation{
+						State: gardencorev1beta1.LastOperationStateFailed,
+					},
+					Gardener: gardencorev1beta1.Gardener{
+						Version: version.Get().GitVersion,
+					},
+				},
+			},
+			BeTrue()),
+	)
+
+	DescribeTable("#IsObservedAtLatestGenerationAndSucceeded",
+		func(shoot *gardencorev1beta1.Shoot, match gomegatypes.GomegaMatcher) {
+			Expect(IsObservedAtLatestGenerationAndSucceeded(shoot)).To(match)
+		},
+
+		Entry("not at observed generation",
+			&gardencorev1beta1.Shoot{
+				ObjectMeta: metav1.ObjectMeta{Generation: 1},
+			},
+			BeFalse()),
+		Entry("last operation state not succeeded",
+			&gardencorev1beta1.Shoot{
+				Status: gardencorev1beta1.ShootStatus{
+					LastOperation: &gardencorev1beta1.LastOperation{
+						State: gardencorev1beta1.LastOperationStateError,
+					},
+				},
+			},
+			BeFalse()),
+		Entry("observed at latest generation and no last operation state",
+			&gardencorev1beta1.Shoot{},
+			BeFalse()),
+		Entry("observed at latest generation and last operation state succeeded",
+			&gardencorev1beta1.Shoot{
+				Status: gardencorev1beta1.ShootStatus{
+					LastOperation: &gardencorev1beta1.LastOperation{
+						State: gardencorev1beta1.LastOperationStateSucceeded,
+					},
+				},
+			},
+			BeTrue()),
+	)
+
+	DescribeTable("#SyncPeriodOfShoot",
+		func(respectSyncPeriodOverwrite bool, defaultMinSyncPeriod time.Duration, shoot *gardencorev1beta1.Shoot, expected time.Duration) {
+			Expect(SyncPeriodOfShoot(respectSyncPeriodOverwrite, defaultMinSyncPeriod, shoot)).To(Equal(expected))
+		},
+
+		Entry("don't respect overwrite",
+			false,
+			1*time.Second,
+			&gardencorev1beta1.Shoot{},
+			1*time.Second),
+		Entry("respect overwrite but no overwrite",
+			true,
+			1*time.Second,
+			&gardencorev1beta1.Shoot{},
+			1*time.Second),
+		Entry("respect overwrite but overwrite invalid",
+			true,
+			1*time.Second,
+			&gardencorev1beta1.Shoot{
+				ObjectMeta: metav1.ObjectMeta{
+					Annotations: map[string]string{v1beta1constants.ShootSyncPeriod: "foo"},
+				},
+			},
+			1*time.Second),
+		Entry("respect overwrite but overwrite too short",
+			true,
+			2*time.Second,
+			&gardencorev1beta1.Shoot{
+				ObjectMeta: metav1.ObjectMeta{
+					Annotations: map[string]string{v1beta1constants.ShootSyncPeriod: (1 * time.Second).String()},
+				},
+			},
+			2*time.Second),
+		Entry("respect overwrite with longer overwrite",
+			true,
+			2*time.Second,
+			&gardencorev1beta1.Shoot{
+				ObjectMeta: metav1.ObjectMeta{
+					Annotations: map[string]string{v1beta1constants.ShootSyncPeriod: (3 * time.Second).String()},
+				},
+			},
+			3*time.Second),
+	)
+
+	Describe("#EffectiveMaintenanceTimeWindow", func() {
+		It("should shorten the end of the time window by 15 minutes", func() {
+			var (
+				begin = utils.NewMaintenanceTime(0, 0, 0)
+				end   = utils.NewMaintenanceTime(1, 0, 0)
+			)
+
+			Expect(EffectiveMaintenanceTimeWindow(utils.NewMaintenanceTimeWindow(begin, end))).
+				To(Equal(utils.NewMaintenanceTimeWindow(begin, utils.NewMaintenanceTime(0, 45, 0))))
+		})
+	})
+
+	DescribeTable("#EffectiveShootMaintenanceTimeWindow",
+		func(shoot *gardencorev1beta1.Shoot, window *utils.MaintenanceTimeWindow) {
+			Expect(EffectiveShootMaintenanceTimeWindow(shoot)).To(Equal(window))
+		},
+
+		Entry("no maintenance section",
+			&gardencorev1beta1.Shoot{},
+			utils.AlwaysTimeWindow),
+		Entry("no time window",
+			&gardencorev1beta1.Shoot{
+				Spec: gardencorev1beta1.ShootSpec{
+					Maintenance: &gardencorev1beta1.Maintenance{},
+				},
+			},
+			utils.AlwaysTimeWindow),
+		Entry("invalid time window",
+			&gardencorev1beta1.Shoot{
+				Spec: gardencorev1beta1.ShootSpec{
+					Maintenance: &gardencorev1beta1.Maintenance{
+						TimeWindow: &gardencorev1beta1.MaintenanceTimeWindow{},
+					},
+				},
+			},
+			utils.AlwaysTimeWindow),
+		Entry("valid time window",
+			&gardencorev1beta1.Shoot{
+				Spec: gardencorev1beta1.ShootSpec{
+					Maintenance: &gardencorev1beta1.Maintenance{
+						TimeWindow: &gardencorev1beta1.MaintenanceTimeWindow{
+							Begin: "010000+0000",
+							End:   "020000+0000",
+						},
+					},
+				},
+			},
+			utils.NewMaintenanceTimeWindow(
+				utils.NewMaintenanceTime(1, 0, 0),
+				utils.NewMaintenanceTime(1, 45, 0))),
+	)
+})
