+++ test/framework/gardener_utils.go
@@ -165,7 +165,7 @@ func (f *GardenerFramework) DeleteShootAndWaitForDeletion(ctx context.Context, s
 // DeleteShoot deletes the test shoot
 func (f *GardenerFramework) DeleteShoot(ctx context.Context, shoot *gardencorev1beta1.Shoot) error {
 	err := retry.UntilTimeout(ctx, 20*time.Second, 5*time.Minute, func(ctx context.Context) (done bool, err error) {
-		err = f.RemoveShootAnnotation(ctx, shoot, common.ShootIgnore)
+		err = f.RemoveShootAnnotation(ctx, shoot, gardencorev1beta1constants.ShootIgnore)
 		if err != nil {
 			return retry.MinorError(err)
 		}
